requirements.txt
fastapi==0.118.0
python-jose==3.5.0
passlib==1.7.4
pydantic==2.11.10
SQLAlchemy==2.0.43
psycopg2-binary==2.9.11
uvicorn==0.29.0
python-dotenv==1.0.1
argon2_cffi==23.1.0
pydantic[email]==2.11.10
gunicorn==22.0.0
pyjwt==2.9.0
python-multipart==0.0.12
bcrypt==4.2.0
